requirements.txt
Flask==2.0.1
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
torch==2.0.0
torchvision==0.15.1
mysql-connector-python==8.0.32

